Add locking scope
A locking scope can be used if multiple applications use the
same redis store for locking.

This is also useful in case tests are run in paralleAdd locking scope

A locking scope can be used if multiple applications use the
same redis store for locking.

This is also useful in case tests are run in parallell.